# documentacion/OpenAPI/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9000/",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/producto/{id}": {
      "get": {
        "tags": [
          "Producto-controller"
        ],
        "summary": "Devuelve un producto",
        "description": "Servicio encargado de retornar un producto con id ingresado por parámetro {id}.",
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Producto-controller"
        ],
        "summary": "Actualiza un producto",
        "description": "Servicio encargado de actualizar un producto con el ID especificado, devolviendo el producto actualizado.",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "Producto-controller"
        ],
        "summary": "Elimina un producto",
        "description": "Servicio encargado de eliminar un producto con el ID especificado.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/producto": {
      "get": {
        "tags": [
          "Producto-controller"
        ],
        "summary": "Devuelve el listado: Productos",
        "description": "Servicio encargado de devolver el listado de productos.",
        "operationId": "getAll_1",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Producto-controller"
        ],
        "summary": "Persiste un producto",
        "description": "Servicio encargado de persistir un producto, retorna el producto con el id asignado.",
        "operationId": "save_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Producto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Producto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}